Added total score.

+++ Main/src/net/ivang/axonix/main/screen/game/actor/NotificationWindow.java
@@ -106,11 +106,11 @@ public class NotificationWindow extends Table {
         levelScoreValue.setText(Integer.toString(score));
     }
 
-    public void setTotalScore(int score) {
-        totalScoreValue.setText(Integer.toString(score));
+    public void setTotalScore(long score) {
+        totalScoreValue.setText(Long.toString(score));
     }
 
-    public void setScores(int levelScore, int totalScore) {
+    public void setScores(int levelScore, long totalScore) {
         setLevelScore(levelScore);
         setTotalScore(totalScore);
     }
